Settings Struct and Config File
I have a vector of 3 string trio's that store the type, name and value
of the data in the config file. I'm hoping this is what you wanted
Charlie. I haven't yet added the functionality to save and load settings
yet or to create that vector of char inputs that you wanted too but that
should be pretty easy. Also I know we want to make the data private too.